rename queue entry to serdes.
