--- notify-server/notify_server/clients/queue_entry.py
@@ -1,61 +1,0 @@
-"""Implementation of the QueueEntry class."""
-from __future__ import annotations
-
-from typing import List
-
-from pydantic import BaseModel
-
-from notify_server.models.event import Event
-
-
-class MalformedFrames(Exception):
-    """Exception raised on badly formed frames."""
-
-    pass
-
-
-class FrameEncodingError(Exception):
-    """Exception raise on encoding error."""
-
-    pass
-
-
-class QueueEntry(BaseModel):
-    """An entry in a send/receive queue."""
-
-    topic: str
-    event: Event
-
-    def to_frames(self) -> List[bytes]:
-        """
-        Create zmq frames from members.
-
-        :raises: FrameEncodingError
-        """
-        try:
-            event_json = self.event.json()
-        except (ValueError, TypeError) as e:
-            # Could not encode event to json.
-            raise FrameEncodingError() from e
-
-        return [
-            bytes(v, 'utf-8') for v in (self.topic, event_json,)
-        ]
-
-    @classmethod
-    def from_frames(cls, frames: List[bytes]) -> QueueEntry:
-        """
-        Create a QueueEntry from a zmq frame.
-
-        The frame must have two entries: a topic, a json serialized Event
-        object.
-
-        :raises: MalformedFrame
-        """
-        try:
-            return cls(
-                topic=frames[0].decode(),
-                event=Event.parse_raw((frames[1]))
-            )
-        except (ValueError, IndexError, AttributeError) as e:
-            raise MalformedFrames() from e
