+++ notify-server/notify_server/clients/subscriber.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 import logging
 import typing
 
-from notify_server.clients.queue_entry import QueueEntry
+from notify_server.clients.serdes import SubscriberEntry, from_frames
 from notify_server.network.connection import create_subscriber, Connection
 
 log = logging.getLogger(__name__)
@@ -35,15 +35,15 @@ class Subscriber:
         """Stop the subscriber task."""
         self._connection.close()
 
-    async def next_event(self) -> QueueEntry:
+    async def next_event(self) -> SubscriberEntry:
         """Get next event."""
         s = await self._connection.recv_multipart()
-        return QueueEntry.from_frames(s)
+        return from_frames(s)
 
     def __aiter__(self) -> 'Subscriber':
         """Create an async iterator."""
         return self
 
-    async def __anext__(self) -> QueueEntry:
+    async def __anext__(self) -> SubscriberEntry:
         """Get next event."""
         return await self.next_event()
